services/api/pyproject.toml
[project]
name = "plerion_api"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "boto3==1.39.5",
    "fastapi>=0.116.1",
    "mangum>=0.19.0",
    "piccolo[postgres]>=1.27.1",
    "pydantic-settings>=2.10.1",
    "uvicorn[standard]>=0.35.0",
]

[dependency-groups]
dev = [
    "boto3-stubs[s3]>=1.39.7",
    "botocore-stubs>=1.38.46",
    "openapi-generator-cli>=7.14.0",
    "ruff>=0.12.3",
    "watchdog>=6.0.0",
]

[tool.pyright]
typeCheckingMode = "strict"

[tool.ruff.lint]
extend-select = ["I"]

[project.scripts]
db = "src.db.tool:main"

[tool.uv]
package = true

[build-system]
requires = ["hatchling>=1.25"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]
